+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 - updated to @applitools/eyes-sdk-core@12.2.4 (from 12.2.0)
 - updated to @applitools/visual-grid-client@15.0.4 (from 15.0.0)
-
+- remove yarn workspaces
 
 ## 2.22.0 - 2020/8/30