+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -6,6 +6,8 @@
 - remove yarn workspaces
 - updated to @applitools/eyes-sdk-core@12.2.6 (from 12.2.5)
 - updated to @applitools/visual-grid-client@15.0.6 (from 15.0.5)
+- updated to @applitools/eyes-sdk-core@12.3.0 (from 12.2.6)
+- updated to @applitools/visual-grid-client@15.0.9 (from 15.0.6)
 
 ## 2.22.0 - 2020/8/30